requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.4
python-dotenv==1.0.0
Werkzeug==2.3.7
Flask-Cors==4.0.0
gunicorn==21.2.0
